@@ -1,10 +1,11 @@
 """Configure pytask."""
+from __future__ import annotations
+
 import os
 from typing import Any
-from typing import Dict
 
-from _pytask.config import hookimpl
-from _pytask.shared import get_first_non_none_value
+from pytask import get_first_non_none_value
+from pytask import hookimpl
 from pytask_parallel.backends import PARALLEL_BACKENDS_DEFAULT
 from pytask_parallel.callbacks import delay_callback
 from pytask_parallel.callbacks import n_workers_callback
@@ -13,9 +14,9 @@ from pytask_parallel.callbacks import parallel_backend_callback
 
 @hookimpl
 def pytask_parse_config(
-    config: Dict[str, Any],
-    config_from_cli: Dict[str, Any],
-    config_from_file: Dict[str, Any],
+    config: dict[str, Any],
+    config_from_cli: dict[str, Any],
+    config_from_file: dict[str, Any],
 ) -> None:
     """Parse the configuration."""
     config["n_workers"] = get_first_non_none_value(
@@ -46,7 +47,7 @@ def pytask_parse_config(
 
 
 @hookimpl
-def pytask_post_parse(config: Dict[str, Any]) -> None:
+def pytask_post_parse(config: dict[str, Any]) -> None:
     """Disable parallelization if debugging is enabled."""
     if config["pdb"] or config["trace"]:
         config["n_workers"] = 1
