@@ -1,6 +1,8 @@
 """Extend the build command."""
+from __future__ import annotations
+
 import click
-from _pytask.config import hookimpl
+from pytask import hookimpl
 from pytask_parallel.backends import PARALLEL_BACKENDS
 from pytask_parallel.backends import PARALLEL_BACKENDS_DEFAULT
 
