@@ -1,22 +1,17 @@
+from __future__ import annotations
+
 import pickle
 import textwrap
+from pathlib import Path
 from time import time
 
-import attr
 import pytest
 from pytask import cli
 from pytask import main
+from pytask import Task
 from pytask_parallel.backends import PARALLEL_BACKENDS
 from pytask_parallel.execute import DefaultBackendNameSpace
 from pytask_parallel.execute import ProcessesNameSpace
-
-
-@attr.s
-class DummyTask:
-    function = attr.ib()
-
-    def execute(self):
-        self.function()
 
 
 class Session:
@@ -116,7 +111,7 @@ def test_pytask_execute_task_w_processes(parallel_backend):
     with pytest.raises(AttributeError):
         pickle.dumps(myfunc)
 
-    task = DummyTask(myfunc)
+    task = Task(base_name="task_example", path=Path(), function=myfunc)
 
     session = Session()
     session.config = {
@@ -265,3 +260,29 @@ def test_rendering_of_tracebacks_with_rich(
     assert "───── Traceback" in result.output
     assert ("───── locals" in result.output) is show_locals
     assert ("[0, 1, 2, 3, 4]" in result.output) is show_locals
+
+
+@pytest.mark.end_to_end
+@pytest.mark.parametrize("parallel_backend", PARALLEL_BACKENDS)
+def test_generators_are_removed_from_depends_on_produces(tmp_path, parallel_backend):
+    """Only works with pytask >=0.1.9."""
+    source = """
+    from pathlib import Path
+    import pytask
+
+    @pytask.mark.parametrize("produces", [
+        ((x for x in ["out.txt", "out_2.txt"]),),
+        ["in.txt"],
+    ])
+    def task_example(produces):
+        produces = {0: produces} if isinstance(produces, Path) else produces
+        for p in produces.values():
+            p.write_text("hihi")
+    """
+    tmp_path.joinpath("task_dummy.py").write_text(textwrap.dedent(source))
+
+    session = main(
+        {"paths": tmp_path, "parallel_backend": parallel_backend, "n_workers": 2}
+    )
+
+    assert session.exit_code == 0
